requirements.txt
beautifulsoup4==4.12.0
boto3==1.35.94
calmjs.parse==1.3.0
celery==5.3.6
channels==4.2.0
channels-redis==4.2.1
chompjs==1.1.9
cryptography>=38.0.3
csscompressor==0.9.5
dj-database-url==0.5.0
django-annoying==0.10.6
django-anymail[amazon-ses]==10.0
django-brotli==0.2.0
django-binary-database-files==1.0.18
django-compressor==4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-heroku-postgresify==0.4.1
django-hosts==5.1
django-htmlmin==0.11.0
django-localflavor==3.1
django_redis==5.2.0
django_simple_history==3.1.1
django-storages==1.12.3
django-static-precompiler==2.1
django_webtest==1.9.10
Django==4.2.16
Faker==13.12.1
flake8==4.0.1
freezegun==1.2.1
google-search-results==2.4.2
hashids==1.3.1
honcho==1.1.0
ipython==8.18.1
kombu==5.3.7
jwt==1.3.1
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
metaphor_python==0.1.10
mock==4.0.3
openai==1.58.1
rollbar==0.16.3
Pillow==10.0.1
piprot==0.9.11
polytester==1.2.3
psycopg2==2.9.10
python-docx==0.8.11
python-pptx==0.6.21
PyMuPDF==1.25.1
PyPDF2==3.0.1
# pyppeteer==1.0.2
pytz==2022.1
rapidfuzz==3.1.1
requests[security]==2.32.4
# scrapy==2.9.0
# scrapy-splash==0.9.0
quill-delta[html]==1.0.3
user-agents==2.2.0
uvicorn[standard]==0.32.1
weasyprint==54.0
wikipedia==1.4.0
whitenoise==6.2.0
pytesseract==0.3.10
pdf2image==1.16.3
docx2python==2.7.3

# Security updates
urllib3==2.2.3
aiohttp==3.10.11
waitress==3.0.1
setuptools==70.0.0